add middleware to highlight links pointing to the current page.
this functionality used to exist but it was lost when converting
the layout handling to middleware.  i add it back in by creating
another layer of middleware to tranform the tree of html.  i'm
not sure if this is proper (maybe this should've  been added into
the existing wrap-layout function?) but i learned a little bit.

first time using clojure zippers, too.